+++ internal/maintenance/reindexer_test.go
@@ -2,6 +2,7 @@ package maintenance
 
 import (
 	"context"
+	"fmt"
 	"testing"
 	"time"
 
@@ -11,6 +12,7 @@ import (
 	"github.com/riverqueue/river/riverdriver"
 	"github.com/riverqueue/river/riverdriver/riverpgxv5"
 	"github.com/riverqueue/river/rivershared/riversharedtest"
+	"github.com/riverqueue/river/rivershared/startstop"
 	"github.com/riverqueue/river/rivershared/startstoptest"
 )
 
@@ -20,8 +22,9 @@ func TestReindexer(t *testing.T) {
 	ctx := context.Background()
 
 	type testBundle struct {
-		exec riverdriver.Executor
-		now  time.Time
+		exec   riverdriver.Executor
+		now    time.Time
+		schema string
 	}
 
 	setup := func(t *testing.T) (*Reindexer, *testBundle) {
@@ -34,7 +37,8 @@ func TestReindexer(t *testing.T) {
 		)
 
 		bundle := &testBundle{
-			exec: riverpgxv5.New(dbPool).GetExecutor(),
+			exec:   riverpgxv5.New(dbPool).GetExecutor(),
+			schema: schema,
 		}
 
 		archetype := riversharedtest.BaseServiceArchetype(t)
@@ -78,20 +82,55 @@ func TestReindexer(t *testing.T) {
 		startstoptest.Stress(ctx, t, svc)
 	})
 
-	t.Run("ReindexesOne", func(t *testing.T) {
+	t.Run("ReindexOneSuccess", func(t *testing.T) {
 		t.Parallel()
 
 		svc, _ := setup(t)
 
+		// Protect against changes in test setup.
+		require.NotEmpty(t, svc.Config.IndexNames)
+
 		for _, indexName := range svc.Config.IndexNames {
-			require.NoError(t, svc.reindexOne(ctx, indexName))
+			didReindex, err := svc.reindexOne(ctx, indexName)
+			require.NoError(t, err)
+			require.True(t, didReindex)
 		}
+	})
+
+	t.Run("ReindexSkippedWithReindexArtifact", func(t *testing.T) {
+		t.Parallel()
+
+		svc, bundle := setup(t)
+
+		requireReindexOne := func(indexName string) bool {
+			didReindex, err := svc.reindexOne(ctx, indexName)
+			require.NoError(t, err)
+			return didReindex
+		}
+
+		indexName := svc.Config.IndexNames[0]
+
+		// With a `_ccnew` index in place, the reindexer refuses to run.
+		require.NoError(t, bundle.exec.Exec(ctx, fmt.Sprintf("CREATE INDEX %s_ccnew ON %s.river_job (id)", indexName, bundle.schema)))
+		require.False(t, requireReindexOne(indexName))
+
+		// With the index dropped again, reindexing can now occur.
+		require.NoError(t, bundle.exec.Exec(ctx, fmt.Sprintf("DROP INDEX %s.%s_ccnew", bundle.schema, indexName)))
+		require.True(t, requireReindexOne(indexName))
+
+		// `_ccold` also prevents reindexing.
+		require.NoError(t, bundle.exec.Exec(ctx, fmt.Sprintf("CREATE INDEX %s_ccold ON %s.river_job (id)", indexName, bundle.schema)))
+		require.False(t, requireReindexOne(indexName))
+
+		// And with `_ccold` dropped, reindexing can proceed.
+		require.NoError(t, bundle.exec.Exec(ctx, fmt.Sprintf("DROP INDEX %s.%s_ccold", bundle.schema, indexName)))
+		require.True(t, requireReindexOne(indexName))
 	})
 
 	t.Run("ReindexesEachIndex", func(t *testing.T) {
 		t.Parallel()
 
-		svc, _ := setup(t)
+		svc, bundle := setup(t)
 
 		svc.Config.IndexNames = []string{
 			"river_job_kind",
@@ -108,6 +147,63 @@ func TestReindexer(t *testing.T) {
 			require.FailNow(t, "Didn't expect reindexing to occur again")
 		case <-time.After(100 * time.Millisecond):
 		}
+
+		// Make sure that no `CONCURRENTLY` artifacts exist after reindexing is
+		// supposed to be done. Postgres creates a new index suffixed with
+		// `_ccnew` before swapping it in as the new index. The existing index
+		// is renamed `_ccold` before being dropped concurrently.
+		//
+		// https://www.postgresql.org/docs/current/sql-reindex.html#SQL-REINDEX-CONCURRENTLY
+		for _, indexName := range svc.Config.IndexNames {
+			for _, reindexArtifactName := range []string{indexName + "_ccnew", indexName + "_ccold"} {
+				indexExists, err := bundle.exec.IndexExists(ctx, &riverdriver.IndexExistsParams{Index: reindexArtifactName, Schema: bundle.schema})
+				require.NoError(t, err)
+				require.False(t, indexExists)
+			}
+		}
+	})
+
+	t.Run("ReindexDeletesArtifactsWhenCancelledWithStop", func(t *testing.T) {
+		t.Parallel()
+
+		svc, bundle := setup(t)
+		svc.skipReindexArtifactCheck = true
+
+		requireIndexExists := func(indexName string) bool {
+			indexExists, err := bundle.exec.IndexExists(ctx, &riverdriver.IndexExistsParams{Index: indexName, Schema: bundle.schema})
+			require.NoError(t, err)
+			return indexExists
+		}
+
+		var (
+			indexName    = svc.Config.IndexNames[0]
+			indexNameNew = indexName + "_ccnew"
+			indexNameOld = indexName + "_ccold"
+		)
+
+		require.NoError(t, bundle.exec.Exec(ctx, fmt.Sprintf("CREATE INDEX %s ON %s.river_job (id)", indexNameNew, bundle.schema)))
+		require.NoError(t, bundle.exec.Exec(ctx, fmt.Sprintf("CREATE INDEX %s ON %s.river_job (id)", indexNameOld, bundle.schema)))
+
+		require.True(t, requireIndexExists(indexNameNew))
+		require.True(t, requireIndexExists(indexNameOld))
+
+		{
+			// Pre-cancel context to simulate a reindexer being stopped while
+			// building a new index.  This requires use of
+			// `skipReindexArtifactCheck` above because checking for reindex
+			// artifacts is the first thing the function does upon entry, and
+			// normally a cancelled context would error on that step first.
+			// Using the flag lets it drop through that to test `REINDEX`.
+			ctx, cancel := context.WithCancelCause(ctx)
+			cancel(startstop.ErrStop)
+
+			didReindex, err := svc.reindexOne(ctx, indexName)
+			require.ErrorIs(t, err, context.Canceled)
+			require.False(t, didReindex)
+		}
+
+		require.False(t, requireIndexExists(indexNameNew))
+		require.False(t, requireIndexExists(indexNameOld))
 	})
 
 	t.Run("StopsImmediately", func(t *testing.T) {